Merged PR 134: Merge 20411_Icon_DBConnectionIssue to master
- Refactored the API Controller to use a DbContextFactory instead of the RenewableContext
- Replaced the RenewableContext with the DbContextFactory in the Mammoth API Controller
- Changed the non prod configs to have the the SendPreviousItemLinkGroup turned on by default (this was a request by the R10 team)

By replacing the shared RenewableContext with the DbContextFactory it makes it so we can create a new context for each request to the database and then dispose of the context immediately once it's done via a using statement. That should eliminate any lasting connections to the database.

Related work items: #20411