Merged PR 141: Added a QueryHandler to to pull whether a message is a Non-Retail Product
Added a QueryHandler to to pull whether a message is a Non-Retail Product message

The API Controller was failing with an exception that the ObjectContext had already been disposed. This was because of the recent changes where we open and then close the context immediately in the commandhandlers and queryhandlers.
Because the context was disposed of the MessageHistoryProcessor would fail when attempting to read the first MessageQueueProduct of that MessageHistory record. This change encapsulates that check in a QueryHandler which attaches the MessageHistory record to a new context and then reads the first record.

Related work items: #20650