Merged PR 151: Merge 20737_APIController_ESBShutDown to master
Fixed a test in the Mammoth API Controller
Change the Icon API Controller to handle errors caused by not being able to log to the event viewer
Changed the default app.config to point to dev instances of Icon DB and the ESB

The Icon API Controller looked like it was failing because of trying to write to the Event Log without proper permissions. By putting a try/catch around it the code can now call the timer.Start() method even if we fail to write to the Event Log. And the Mammoth controller was failing because it only has a single instance of the ESB Producer. So changing it's lifestyle to not be a singleton allowed the controller to run after the failure.

Related work items: #20737