Merged PR 224: Merge 20862_IRMADb_InforItemID to master
This is the data surgery / data pop scripts to add ItemID in the IRMA Databases.

Tom - I'm not sure how you want to handle the schema changes for this kind of deal.  Should this actually go into the PreDeploy folder instead and then I'll need to update the dbo.ValidatedScanCode table?

Related work items: #21775, #20862