Merged PR 250: Merge 21662_SSRS_ItemChangesReport to master
Changed report so that there would be one row per item that was updated from Icon.  The report should:
- Highlight the field in red that changed

I did this to check a couple of things:
- make sure we compare any change within the date range to the state of the item before the date range
- if there was no item change history row before the date range, compare the rows within the date range.

It's not the prettiest, but QA passed it.  I wasn't thinking and should have submitted this before.  I guess I was making sure the testing passed because I didn't have Infor Test access and needed their help to test it anyway.

Related work items: #21662