Merged PR 304: Fixed creating messages from Icon to R10
I missed actually passing the command handler to the service. So this is basically fixing that with a couple of small tweaks.

Fixed injecting the GenerateHierarchyClassMessagesCommandHandler into the DeleteHierarchyClassService
Updated the tests to not look for queued Events
Added a catch all ApplicationError for when an unexpected error occurs
Added archiving records with an unexpected error code when it occurs