Merged PR 321: Merge 21013_dashboard_deployment to master
Getting rid of Task Scheduler NuGet package entirely because the 3rd party code was flagged for malware. We will just completely stop using/tracking scheduled tasks in the dashboard now since we only have one remaining and we will convert it to a service soon. That way we can get rid of the Task Scheduler package entirely instead of trying to go back to an earlier version or something like that.

Related work items: #21013, #22524