Merged PR 337: Merge 22693_dashboard_build_failure to master
Icon Dashboard: removing bad library reference and updating web.config(s). The build definition was set to use the Release configuration for the dashboard but there is no dashboard in prod so for now I set the "Release" transformed web.config to point to Test. Also a test project had an old reference

Related work items: #21013, #22693