Merged PR 382: Merge 23027_GPM_PriceListenerDeploy to master
Changed a few things in order for the listener to work while deployed.  Also updated some of the file structure, deleted some test processes, and tweaked some global variables.  I didn't change any of the structure of the actual code other than added a transition for the 'AddPriceCommand' to the 'End' of the group which was missing.

Related work items: #23027