Merged PR 383: Merge 22835_InforNewItemService_ItemRefresh to master
I added code to the new item service to not send an item message to Infor if the item exists in Icon.

It pulls the IDs from Icon based on the scan code by using a new stored procedure.
Then it generates events for the ones that have IDs in Icon and send messages for the ones that don't have Icon IDs.

Related work items: #22835