Merged PR 415: Updated the data type of Price column in the Mammoth Price tables to
Updated the data type of Price column in the Mammoth Price tables to decimal(9,2).

Since all the tables below except for the [gpm].[PriceHistory] table are deployed to QA and Prod without data populated, we can simply drop these tables in QA and Prod and re-run the table create the scripts.

Related work items: #23359