Merged PR 422: Merge 21036_SharedCode_ConfirmationBODEsbService to master
I added a new ESB service that builds ConfirmBOD message which we will send to ESB and then to Infor. ConfirmBODs have error information that Infor consumes. It has Error properties and stores the original message in a Base64 encoded string wrapped in a CDATA section.

This library is going to be consumed by current listeners and will return errors to Infor.

Related work items: #21036