Merged PR 469: Merge 23879_icon_dashboard_deploy to master
Adding the .ps1 file so that it will be accessible to the build. Once the powershell script is included in the build, I will be able to add tasks in the Release to backup & restore the dashboard data file. (The other (.cs) changes are just a minor refactoring, adding a base controller class & cosmetic changes + one bug fix.)

Related work items: #23879