Merged PR 470: Merge 23935_InforNewItemService_NotSendingMessagesBug to master
Added a Serializer that doesn't set the encoding of the XML message
Updated the New Item Service with the new Serializer so that it doesn't set the encoding to UTF-8 which causes the message to get rejected when trying to save to the database because SQL Server is expecting a UTF-16 message. By not specifying the encoding the message can be saved to the database and sent to the ESB and both systems consume it.

Related work items: #23935