Merged PR 466: Set future REG end dates to NULL.
Set future REG end dates to NULL.
Changed the filter for future regs to pull all REG prices because a user can make a REG change when adding a TPR change. So there is no real indicator to note if a PriceBatchDetail only has a TPR change and if we should exclude the REG from it.
Changed the PRICE_TYPE to REG on Future Long Running TPRs that are becoming REGs

Related work items: #23878