Merged PR 479: Merge 23624_IRMADB_DelieverPriceFromMammothPriceQueueToIRMADB to master
Fixed a bug identified in the testing. POSPrice and POSSale_Price will be populated together with Price and Sale_Price fields on the Price table. Also, the Start Date will be poupulated together with Price on the PriceBatchDetail record for a new item. This way, the new item chanage request can be batched and pushed.