Merged PR 522: Merge 20107_MammothLogging_NugPkg_Take2 to master
Looks like putting `$(SolutionDir)` in the csproj file prevented the build from working. I switched it back to the `..\` and it finally ran.  I'm going to change the nuget package build to point back to master after this is approved.  I had ended up switching the nuget build for Mammoth Logging to this branch temporarily.