Merged PR 579: Grouping by BU and ItemID so that we can compare prices properly.  Before we
Grouping by BU and ItemID so that we can compare prices properly.  Before we grouped by only BU and other itemIDs were getting left out of the results.

If anyone else has a better way to get current REG price let me know.  But I think this allows for not having to iterate through the priceGroup twice for the REGs.

Related work items: #24605