Merged PR 621: Merge 24518_MammothWebSupport_RefreshGpmPricesFromMammoth to master
Added a new page to the Mammoth Web Support site that allows someone to generate a price message and send it down to Just in Time systems which are R10 and IRMA.
Created a PriceRefresh service which validates the input from the user and also sends the messages to the ESB.
Created a couple of factories to send the generate a message producer to the ESB and also a message builder for the prices.
Added Unit Tests for all the new code.

Related work items: #24518, #24913