Merged PR 693: Merge 24741_PrimeAffinityController_CreateController to master
I created a Prime Affinity Controller that reads prices from Mammoth and sends a AddOrUpdate or Delete Prime Affinity PSG to R10 based on the prices. This is only for TPRs.
If the price is in ISS, SAL, or FRZ and it is starting today then we will send an AddOrUpdate PSG.
If the price is in ISS, SAL, or FRZ and it is ending anytime yesterday to the beginning of today then we will send a Delete PSG.
If the price is is not in ISS, SAL, or FRZ and it is starting today then we will send a Delete PSG.

I also made a couple of changes to the JobSchedule table that we use for the TIBCO services so that this controller could use it to schedule when to send prices. I added a 'failed' status in case something goes really wrong and I removed the requirement for a DestinationQueueName. I updated the TIBCO job scheduler to only pick up jobs that have a DestinationQueueName.

I also made a library for the Prime Affinity code that sends PSGs to the ESB because we will need to basically do the same thing in the Mammoth Web API.

Related work items: #24741