Merged PR 747: Updated the IRMA Price Listener stored procedure to Update the POSPrice,
Updated the IRMA Price Listener stored procedure to Update the POSPrice, POSSale_Price, and PricingMethod_ID on the PriceBatchDetail table. Also updating the PricingMethod_ID on the Price table.

This is so we pass the correct price in the Scale Push during the interim between GPM go live and OnePlum go live.

The PricingMethod_ID is defaulted to 'Regular Promo' because it exists in all regions and is the most used in each region. It also seems to be a price attribute that is ignored by the regions but still used in the Scale Push and other parts of IRMA Client so we can't get rid of it.

Related work items: #25327