Merged PR 775: Added ESL1Plum string to nonReceivingSystemName app setting.
Added ESL1Plum string to nonReceivingSystemName app setting based on the below criteria.  In summary, they will be subscribed to the main Item queue which is where all ItemLocale and Price messages are going to both out of Icon and Mammoth.  They need just-in-time prices, and near real-time ItemLocale messages.  So we are excluding from near real-time price messages and just-in-time ItemLocale messages.

- Add ESL1Plum exclusion to the nonReceivingSysName JMS Header for ItemLocale messages coming out of IRMA via Icon. (i.e. Icon ItemLocale API Controller)
- Add ESL1Plum exclusion to the nonReceivingSysName JMS Header for Price messages coming out of IRMA via Mammoth (i.e. Mammoth Price API Controller)

Related work items: #25830