Merged PR 782: Cancel All Sales Rollback sends Mammoth Price event
Made it so that when rolling back a CancelAllSales PBD we would insert a Mammoth Price Event so that the current price and sale information are sent to Mammoth.

Originally we would just resend the CancelAllSales PBD which wouldn't change Mammoth back to having the sale info.

Related work items: #26247