Merged PR 876: 26905_MammothWebApi_FixGeneratingConnectionsToEsb
We have an issue in Prod where the Mammoth Web API is not disposing of it's connections to the ESB. So I'm changing the ESB connection code to have only a single static Connection and ConnectionFactory and I created a new producer that is returned from that Connection.
I've put this code only in the Prime Affinity Library because adding this to the other ESB projects would take more time. So I think we can refactor and update that code later.
I also created a recurring background job in the Mammoth Web API that every 3 seconds will check its ESB connection and attempt to reconnect to the ESB if it ever becomes disconnected

Related work items: #26905