Merged PR 891: Fix Ad Hoc Issue in Mammoth Prime Affinity Controller
Added an Instance ID to the JobSchedule table so that the jobs don't step on each other when we run them and there are multiple instances.
Made the start button on the Web Support not start the job if the job is already running.
Fixed logging issue on the Mammoth Web API.
Fixed an issue on the Prime Affinity Controller where the NextScheduledRunTimeUtc would get increased if the job was run AddHoc

Related work items: #26957