Merged PR 915: Mammoth Web Support PSG fixes and merge of code from PSG release branch
Applied the same changes that we made for one of the bug fixes to this branch on the Prime Affinity Library so that the connection issue is fixed.
Updated the projects with the new version of the Prime Affinity library.
Changed the JobSheduleViewModel to have properties instead of fields so that the compile time warnings in the project would go away.
Updated most of the NuGet packages in each project except for jQuery and Bootstrap because the CSS seems to break if I mess with those.
Added an assembly reference to System.Runtime 4.0.0.0 in the Mammoth Web Support web.config because when deployed to our servers the website breaks and this fixes it.
Removed unused binding redirects in web.config files.

Related work items: #27204