Merged PR 1018: Bug 27752: Mammoth - Database Error when the percent off value received is greater than or equal to 10 for a Rewards by percentage message
Increases the size of the SQL type used for storing PercentOff in Mammoth DB. Changing it from DECIMAL(3,2) to DECIMAL(5,2) will allow a range of values from 0.00 to 999.99 to be stored. DECIMAL(3,2) only allows between 0.00 and 9.99. The ESB Schema allows values up to 100 (3 digits before the decimal) so the old (3,2) format would fail on any value 10 or above.

Related work items: #27752