Merged PR 1204: Consolidating deployment config files to a single file under renamed 'Config'
Consolidating deployment config files to a single file under renamed 'Config' folder that must be manually edited when we deploy to QA (file staying in Git is correct for Perf and Prod).  Deleting QAF.old folder.  Deleting .XML files under QAF folder (these are the extra copies of deploy config that are special to QA, but we will manually setup for QA until a different solution or process is setup).