Merged PR 1277: Added table for Gpm / 765 conversion for all the prices to be cancelled.  Added
Added table for Gpm / 765 conversion for all the prices to be cancelled.  Added two scripts. One for Cancelling the TPRs and one for updating the IRMA Price table to REG for the long running TPRs turning to REG

For context, the flow of how we will convert all new regions to GPM (with 765 turned on) will be:
1. Receive a flat file that will have all the TPRs that need to be cancelled in IRMA
2. Populate the IconPOSPushPublish table (I use the `Replenishment_POSPush_PopulateNonBatchIconPOSPushStaging` for reference).
3. POS Push Controller (Pusheen) will pick up those rows and messages will eventually get generated from the Api Controller (in Icon).
4. Update IRMA with all the REG prices that used to be "TPRs" in IRMA (price types like DIS, EDV, CLR, etc)

Related work items: #29333, #29953