Merged PR 53: Reinsert into stage table with SYSDATETIME
When exception is caught, re-insert the records with the current datetime instead of the original datetime.
Doing this for both
- esl.GetEslAttributes
- eplum.GetEPlumAttributes

NOTE:  Column is `InsertDateUtc` however, the DDS team inserts and checks for the current datetime which happens to be central.  This is unfortunate but because they do it that way we're going to accommodate and do it the same way.

Related work items: #10814