Merged PR 304: Updated NonReceivingSystem configs for all environments so that manual changes
Updated NonReceivingSystem configs for all environments so that manual changes in PRD do not get overwritten.  The big thing is that we made changes to exclude AMZ from certain messages and this is to keep those in line.

What AMZ should receive:
- Mammoth ItemLocale Messages
- Mammoth and GPM Price Messages
- Icon Global Item Messages

What they should not get:
- Icon ItemLocale data
- Icon Price data
- Prime Psg data

Related work items: #19587