Merged PR 773: Only add Linked Item Groups if there is a LinkedScanCode or PreviousLinkedScanCode because if the groups element is empty then it will break ESB validation
Updated the API Controller to only add Linked Item Groups if there is a LinkedScanCode or PreviousLinkedScanCode because if the groups element is empty then it will break ESB validation.

Also made it so non Deposit and Fee PreviousLinkedScanCodes will be added as delete groups.

Updated unit tests as well.

Related work items: #22632