Merged PR 778: Exclude Unauthorized and Discontinued Items from IRMAAvgCostsToChefTec DM script
build up tables for each item/store  and item/store/primary vendor showing authorization and discontinued states
if a given item is authed or not discontinued in at least 1 open store for the region, include it in the results.
otherwise exclude it.

output for this script is at the item level. not the item/store level.

Related work items: #22345