Merged PR 897: Fixed the file connections to use the package parameters for their file path
We found a bug with the current package where the file paths are basically hard-coded. Fixed it up with this.

Fixed the file connections to use the package parameters for their file path so that you can pass those parameters to the package to dictate where the files are dropped

Related work items: #14578