Merged PR 3996: WFM Mobile Handle 128 barcodes
Based on the information and scan codes available this is limit of what I was able to test with. It may not match every scan code available and have to be reworked when we have more scan codes available later in the project.

- Added ScanCodeProcessor to translate scan codes scanned from the device into scan codes that the apps and Whole Foods would recognize
- Added sending of Symbology to the ScanCodeProcessor
- Updated the .npmc file to point to the new location of the @wfm\mobile npm packages
Removed the BarcodeScanner.scanHandler = () => { }; from the app because this is now a private field on the BarcodeScanner with the latest update. I don't think this was needed in the first place after reviewing the code

Related work items: #32961