Merged PR 4661: 36612 36467 36464 Bug fixes for WFM Mobile Hypercare
Fixed Shrink screen to take SoldByWeight into consideration.
Consolidated toast notifications throughout the application. Changed the default auto close to 3 seconds and moved the toast messages to the top of the app instead of at the bottom which was blocking buttons.
Updated the InvoiceData screen to allow a user to close an order if they choose to use an order type of None.

Related work items: #36464, #36467, #36612