Merged PR 8071: Updated IRMA mobile the way to determine old/new style of token and replace % with apostrophe in user token
There are two issues in the current WFM mobile SSO authentication process:

The 1st issue is that if a user's name contain an apostrophe, the token assembled in the WFM mobile cannot be passed down to client apps like IRMA mobile and OOS mobile without error. The WFM mobile code has been updated to replace the apostrophe with '%' before we send the user info to the client apps, And IRMA mobile and OOS have been updated to reverse % into apostrophe when parsing the user token.

The 2nd issue is that currently the code in IRMA mobile and OOS try to parse the token as a json string first (the way we used to handle the old style user token). If that fails, it will try to handle the token as in the new token style. For some reason, if certain AD attributes of a user are missing value, the new token can be treated as old style token.

IRMA mobile and OOS have been updated to just check if certain string that should only be available in the new token exists to determine if it's a new token or an old token.

Related work items: #47886