Merged PR 13089: Implement EmergencyPrice service
Implement EmergencyPrice service. This service is responsible for sending emergency Prices (emergency prices are prices that are active as soon as the change was made) immediately to JustInTime consumers.

Data Flow: (The flow is similar to ActivePrice service PR#13069)

1. Triggered periodically to check emergency prices.
1. If prices exist,
   1. Dequeue n number of Prices from the table.
   1. Create a single MammothPrices xml message that contains all the dequeued MammothPrice canonical objects within it.
   1. Send to ESB.
   1. If failed,
      1. Insert back to emergency table.
      1. Retry if attempts left, or wait for next trigger and exit directly.
1. Check if prices exist again, before exiting.

`EmergencyPriceMessageProcessor.cs` would be a good starting point to understand the flow.

Related work items: #51280