Merged PR 13121: Grant permissions to MammothRole for GPM process
Grant permissions to `MammothRole` for GPM process. This role is used in the Azure deployment pipeline for the new re-written .NET service for GPM.

These permissions are same as what was granted to `TibcoRole` role.

### Revision change
Added grants at the object level instead of using a pop script as per discussion with Tom (DBA) (tom.lux@wholefoods.com)

Related work items: #51280