Merged PR 13315: Create AWS library
Create AWS library.

As of now this has S3 facade and SNS facade.

The idea of creating this library is to automatically create AWS clients that work using access key and secret key. So, consumers of this library, will use `S3Facade` instead of creating `AmazonS3Client` on their own.

Related work items: #51289