Merged PR 13437: Keep polling inventory events till empty
Keep polling inventory events till empty. After the Tibco apps were stopped for RM region, it is observed that there are many pending events.

The current logic of Inventory Producer service polls 100 events every 5 minutes and stops for next 5 minutes.

This new logic will keep polling events, if events were polled in the previous run, without waiting for next 5 minutes. Only when no events were polled, will the current run stop. **This logic was already present in the Tibco app.** It was intentionally not added in the new service, to have clear periods of no processing (during which service can be stopped safely). The rate at which events are produced is not known. But seeing the number of pending events, it is better to implement the Tibco behaviour.

As a result, reduced the interval at which we check if it safe to stop the service (from 30 seconds to 15 seconds).

Additionally, increased the `DequeueMaxRecords` property to 300. The more number of records we dequeue in one go, the less number of duplicates there will be. This also matches the Tibco app configuration. The count was kept lower previously, because the number of hosts for this windows service is more than the number of Tibco instances.

Related work items: #51252