Merged PR 13516: Enable conditional safe stop in Wfm.Aws
Enable conditional safe stop in Wfm.Aws. The default value would be false.

Some applications may not want the safe stop behaviour (as with this way, the only way to stop application during a traffic period would be force quitting it). Hence, add a flag to check if the application explicitly wants this.

Related work items: #51289