Merged PR 13635: Add support for row lock in JobScheduler to prevent duplicates
Add support for row lock in JobScheduler to prevent duplicates.

The job scheduler logic works by selecting all the jobs that are scheduled, then sends these to downstream, then updates the table with next scheduled time.

Before one instance runs and updates the next scheduled time, another parallel instance may poll the same scheduled jobs and send to downstream resulting in duplicates.

This works in the Tibco app because *Tibco app runs from only one instance*.

To handle this, without compromising on the instance count, introduced the logic to lock the row. *This is a new functionality that is not there in the Tibco app.*

Only one instance will be able to lock the row. After pulling and processing the scheduled jobs, the lock will be released. To prevent an infinite lock (in case an instance went down after locking but before releasing the lock), used the DB Column `NextScheduledDateTimeUtc`. A lock will be overridden for a job, if that job didn't execute in the previous service run even after being scheduled. The lock is based on the value of the `InstanceId` and `NextScheduledDateTimeUtc` column in the JobScheduler table.

This `InstanceId` was an unused column. So, we can use it.

### Still a rare chance of duplicate
There's still a rare chance of duplicate, when one instance failed to release the lock, the other instances will all try to override and if this happens nearly at the same time, all of them may be able to acquire the lock (the lock would actually be owned by the last instance to do so).

Querying for scheduled jobs still uses `InstanceId`, so, in case an instance acquired a lock that got overridden by another instance, before the query could happen, we are still duplicate safe. However, if an instance acquired the lock, performed the query, and then another instance overrode the lock, only then will duplicate happen.

Rare duplicates are fine and it is not expected that duplicates in general will cause a problem.

This is what can be achieved with the current design of Job Scheduler (which was built on the idea that it will run from only one instance).

Related work items: #51286