Merged PR 13661: Stop using ESB for InventoryProducer service
In the current state, messages are sent to both ESB and WfmDataVayuService. Post this change, messages will only be sent to WfmDataVayuService.

It is safe to make this change as all the consumers via ESB (WfmBusinessConstraintLambda, WfmBiInventoryFeedLambda, WfmBIPurchaseOrderFeedLambda, WFMPurchaseOrderManagerService, SnowAvailabilitySignalHandlerLambda and Everclear) have migrated to WfmDataVayuService.

*WFMPurchaseOrderManagerService* is not migrated to DVS yet. It will be migrated before this change gets deployed to Prod.

Changes:
* Removed dependency on `Icon.Esb` library.
* Cleanup us any usages of this library.
* Cleaned up ESB related configs.

Related work items: #51252