Merged PR 13846: Use batch insert instead of transaction for Irma Listener
Use batch insert instead of transaction for inserting multiple items in DB for Irma Listener. There are latency issues in the current application and I suspect it is due to this.
There is no change in business logic, only the way we are interacting with the DB.

Additionally, fixed file based logging.

Tested by deploying in QA FL region with price count as 2.

**Note:** Whether this will solve the latency issue or not can only be verified in Prod, as there is no way to do a performance test in a non prod environment.

Related work items: #51280